FilesizesServletComponent moved into BagsServletComponent
